cynapse-platform/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pdfplumber
pinecone
google-generativeai
requests
python-multipart
pydantic